Revert "#1. Fix Forward-LSTM crash bug #2. Improve encoding performance by SIMD instructions"
This reverts commit 1a3070ce9cf5189b66b912782e50d6a78644b50c.

+++ RNNSharp/MathUtil.cs
@@ -7,7 +7,7 @@ namespace RNNSharp
 {
     class MathUtil
     {
-        public static int GetMaxProbIndex(float [] array)
+        public static int GetMaxProbIndex(double [] array)
         {
             int dim = array.Length;
             double maxValue = array[0];
